+++ src/Client/ReleaseNotes.fs
@@ -1,10 +1,14 @@
 module internal ReleaseNotes
 
-let Version = "0.0.1"
+let Version = "0.0.2"
 
 let IsPrerelease = false
 
 let Notes = """
-### 0.0.1 - 2017-11-24
+### 0.0.2 - 2017-12-07
+* Cleanup, Renaming
+* better handling of MIDI events
+
+### 0.0.1 - 2017-12-06
 * Initial release
 """