+++ src/Client/ReleaseNotes.fs
@@ -8,6 +8,8 @@ let Notes = """
 ### 0.0.2 - 2017-12-07
 * Cleanup, Renaming
 * better handling of MIDI events
+* VS Code tasks for build, run and release
+* GitHub banner
 
 ### 0.0.1 - 2017-12-06
 * Initial release