+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ### 0.0.2 - 2017-12-07
 * Cleanup, Renaming
 * better handling of MIDI events
+* VS Code tasks for build, run and release
 
 ### 0.0.1 - 2017-12-06
 * Initial release