+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 * Cleanup, Renaming
 * better handling of MIDI events
 * VS Code tasks for build, run and release
+* GitHub banner
 
 ### 0.0.1 - 2017-12-06
 * Initial release