+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * better handling of MIDI events
 * VS Code tasks for build, run and release
 * GitHub banner
+* Shareable link
 
 ### 0.0.1 - 2017-12-06
 * Initial release