+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+### 0.0.3 - 2017-12-08
+* Fix accidentally selection on slider movement
+
 ### 0.0.2 - 2017-12-07
 * Cleanup, Renaming
 * better handling of MIDI events